requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
distlib==0.3.1
Django==3.2
djangorestframework==3.12.4
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
grpcio==1.37.0
grpcio-tools==1.37.0
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.24
pipenv==2020.11.15
protobuf==3.15.8
pylint==2.7.1
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.13
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
